Use Resolver API in go-offline for plugins resolving
- use the same method for generating plugin list as in resolve-plugin
- move parameter where is used
- move some method where is used
- cleanups
